test(json-crdt): 💍 update tests after refactor
