+++ src/json-crdt/__tests__/scenario.editing-deleted-part.spec.ts
@@ -3,8 +3,8 @@ import {Encoder as EncoderBinary} from '../codec/binary/Encoder';
 import {Decoder as DecoderBinary} from '../codec/binary/Decoder';
 import {Encoder as EncoderCompact} from '../codec/compact/Encoder';
 import {Decoder as DecoderCompact} from '../codec/compact/Decoder';
-import {LogicalEncoder as EncoderCompactBinary} from '../codec/compact-binary/LogicalEncoder';
-import {LogicalDecoder as DecoderCompactBinary} from '../codec/compact-binary/LogicalDecoder';
+import {Encoder as EncoderCompactBinary} from '../codec/compact-binary/Encoder';
+import {Decoder as DecoderCompactBinary} from '../codec/compact-binary/Decoder';
 import {Encoder as EncoderJson} from '../codec/json/Encoder';
 import {Decoder as DecoderJson} from '../codec/json/Decoder';
 import {encode as encodePatchBinary} from '../../json-crdt-patch/codec/binary/encode';
